requirements.txt
Flask==3.0.0
gunicorn==21.2.0
redis==5.0.1
psycopg2-binary==2.9.9
boto3==1.34.0
botocore==1.34.0
selenium==4.16.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1

